+++ packages/eslint-plugin/src/tests/index.ts
@@ -1,0 +1,2 @@
+import './atom-postfix-rule.test'
+import './rule.test'