+++ packages/jsx/src/index.ts
@@ -1,14 +1,23 @@
-import { createCtx, Ctx, Fn, isAtom, Rec, Unsubscribe } from '@reatom/core'
+import {
+  atom,
+  createCtx,
+  Ctx,
+  Fn,
+  isAtom,
+  Rec,
+  Unsubscribe,
+} from '@reatom/core'
 import type { JSX } from './jsx'
 import { noop } from '@reatom/utils'
-declare type JSXElement = JSX.ElementType
+import { parseAtoms } from '@reatom/lens'
+declare type JSXElement = JSX.Element
 export type { JSXElement, JSX }
 
 type ElementTag = keyof JSX.HTMLElementTags | keyof JSX.SVGElementTags
 
-export type Component<Props = unknown> = (props: Props) => JSXElement
+export type Component<Props> = (props: Props) => JSXElement
 
-export type ComponentLike = ElementTag | Component
+export type ComponentLike = ElementTag | Component<any>
 
 export type InferProps<T extends ComponentLike> =
   T extends keyof JSX.HTMLElementTags
@@ -20,7 +29,7 @@ export type InferProps<T extends ComponentLike> =
     : never
 
 export const reatomJsx = (ctx: Ctx) => {
-  let unsubscribesMap = new WeakMap<HTMLElement, Array<Fn>>()
+  let unsubscribesMap = new WeakMap<Element, Array<Fn>>()
 
   let unlink = (parent: any, un: Unsubscribe) => {
     Promise.resolve().then(() => {
@@ -31,6 +40,46 @@ export const reatomJsx = (ctx: Ctx) => {
         }
       }
     })
+  }
+
+  let render = (parent: Element, children: JSXElement[]) => {
+    // TODO support Atom<Array>
+    let walk = (child: any) => {
+      if (Array.isArray(child)) {
+        child.forEach(walk)
+        return
+      }
+
+      if (isAtom(child)) {
+        let innerChild: ChildNode = document.createTextNode('')
+        var un: Unsubscribe | undefined = ctx.subscribe(
+          atom((ctx) => parseAtoms(ctx, child)),
+          (v) => {
+            if (un && !innerChild.isConnected) un()
+            else {
+              if (v instanceof Element) {
+                let list = unsubscribesMap.get(v)
+                if (!list) unsubscribesMap.set(v, (list = []))
+
+                if (un) parent.replaceChild(v, innerChild)
+                innerChild = v
+              } else {
+                innerChild.textContent = v
+              }
+            }
+          },
+        )
+        unlink(parent, un)
+        parent.appendChild(innerChild)
+        return
+      }
+
+      parent.appendChild(
+        child?.nodeType ? child : document.createTextNode(String(child)),
+      )
+    }
+
+    children.forEach(walk)
   }
 
   let h = ((tag: any, props: Rec, ...children: any[]) => {
@@ -64,36 +113,7 @@ export const reatomJsx = (ctx: Ctx) => {
       }
     }
 
-    let walk = (child: any) => {
-      if (Array.isArray(child)) child.forEach(walk)
-      else {
-        if (isAtom(child)) {
-          let innerChild = document.createTextNode('') as ChildNode
-          var un: undefined | Unsubscribe = ctx.subscribe(child, (v) => {
-            if (un && !innerChild.isConnected) un()
-            else {
-              if (v instanceof HTMLElement) {
-                let list = unsubscribesMap.get(v)
-                if (!list) unsubscribesMap.set(v, (list = []))
-
-                if (un) element.replaceChild(v, innerChild)
-                innerChild = v
-              } else {
-                innerChild.textContent = v
-              }
-            }
-          })
-          unlink(element, un)
-          element.appendChild(innerChild)
-        } else {
-          element.appendChild(
-            child?.nodeType ? child : document.createTextNode(String(child)),
-          )
-        }
-      }
-    }
-
-    children.forEach(walk)
+    render(element, children)
 
     return element
   }) as <T extends ComponentLike>(
@@ -105,8 +125,8 @@ export const reatomJsx = (ctx: Ctx) => {
   /** Fragment */
   let hf = noop
 
-  let mount = (target: Element, child: Element) => {
-    target.appendChild(child)
+  let mount = (target: Element, child: JSXElement) => {
+    render(target, [child])
 
     new MutationObserver((mutationsList) => {
       for (const mutation of mutationsList) {