+++ packages/logger/src/rld.tsx
@@ -1,0 +1,5 @@
+import { Ctx } from '@reatom/core'
+
+export function Rld({ app }: { app: Ctx }) {
+  return <div>Hello world</div>
+}