+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [3.6.0](https://github.com/artalar/reatom/compare/utils-v3.5.0...utils-v3.6.0) (2023-10-09)
+
+
+### Features
+
+* **jsx:** nested atoms, better types ([#652](https://github.com/artalar/reatom/issues/652)) ([3ceae78](https://github.com/artalar/reatom/commit/3ceae788da52ff40a561ce5b2fc5371475fb7d7c))
+* **lens:** add match ([#646](https://github.com/artalar/reatom/issues/646)) ([0f2a768](https://github.com/artalar/reatom/commit/0f2a7685dd797cac4c9fc882a8e24bc31f9503a6))
+
 ## [3.5.0](https://github.com/artalar/reatom/compare/utils-v3.4.0...utils-v3.5.0) (2023-09-13)